+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.1.0
+
+1. Changed the standard Android logging API to **SLF4J Logging API**
+
+# 1.0.1
+
+1. Fixed **NullPointerException** in **MovingParams**
+
 # 1.0.0
 
 1. The first release! Everything is new.